// lib/structures/League/Entry.d.ts
import { APILeagueEntry } from "../../interfaces/League";
export declare class LeagueEntry {
    leagueId: string;
    queueType: string;
    tier: string;
    rank: string;
    summonerId: string;
    summonerName: string;
    leaguePoints: number;
    wins: number;
    losses: number;
    veteran: boolean;
    inactive: boolean;
    freshBlood: boolean;
    hotStreak: boolean;
    constructor(data: APILeagueEntry);
}